update docker-compose and fixing errors

+++ challenge/python-consumer-prev/main.py
@@ -17,8 +17,8 @@ consumer = Consumer({
 
 def my_on_assign(konsumer, partitions):
     "On partition assignment move the offsets to 5 min ago"
-    # Calculate the time 5 minutes ago
-    timestamp = (time.time() - (5 * 60)) * 1000
+    # Calculate the time 5 minutes ago (must be int for offset)
+    timestamp = int((time.time() - (5 * 60)) * 1000)
 
     # Create topic+partitions with timestamps in the TopicPartition.offset field
     for part in partitions:
